Tricky Mirror Test

+++ Tools/TrickyToolsWindow.cs
@@ -1869,7 +1869,82 @@ namespace SSXMultiTool
                         TrickyScaler.ProjectScaler(LoadPath.Substring(0, LoadPath.Length - 4), InputPath.Substring(0, InputPath.Length - 4), Scale);
                     }
 
-                    if(Directory.Exists(Application.StartupPath + "\\TempExtract"))
+                    if (Directory.Exists(Application.StartupPath + "\\TempExtract"))
+                    {
+                        Directory.Delete(Application.StartupPath + "\\TempExtract", true);
+                    }
+                }
+            }
+        }
+
+        private void button4_Click(object sender, EventArgs e)
+        {
+            OpenFileDialog openFileDialog = new OpenFileDialog
+            {
+                Filter = "Map File (*.map,*.big)|*.map;*.big|All files (*.*)|*.*",
+                FilterIndex = 1,
+                RestoreDirectory = false
+            };
+            if (openFileDialog.ShowDialog() == DialogResult.OK)
+            {
+
+                SaveFileDialog openFileDialog1 = new SaveFileDialog
+                {
+                    Filter = "Map File (*.map,*.big)|*.map;*.big|All files (*.*)|*.*",
+                    FilterIndex = 1,
+                    RestoreDirectory = false
+                };
+                if (openFileDialog1.ShowDialog() == DialogResult.OK)
+                {
+                    string LoadPath = openFileDialog.FileName;
+                    string InputPath = openFileDialog1.FileName;
+                    string BuildPath = openFileDialog1.FileName;
+                    bool LoadBig = false;
+
+                    if (LoadPath.ToLower().Contains(".big"))
+                    {
+                        BigHandler bigHandler = new BigHandler();
+                        bigHandler.LoadBig(openFileDialog.FileName);
+                        bigHandler.ExtractBig(Application.StartupPath + "\\TempExtract");
+                        string[] strings = Directory.GetFiles(Application.StartupPath + "\\TempExtract", "*.map", SearchOption.AllDirectories);
+                        if (strings.Length == 0)
+                        {
+                            MessageBox.Show("Missing Level Files");
+                            return;
+                        }
+                        LoadPath = strings[0];
+                        if (BuildPath.Contains(".big"))
+                        {
+                            InputPath = strings[0];
+                        }
+                        LoadBig = true;
+                    }
+
+                    float Scale = 0.6f;
+
+                    if (BuildPath.Contains(".big"))
+                    {
+                        if (!LoadBig)
+                        {
+                            Directory.CreateDirectory(Application.StartupPath + "\\TempExtract");
+                            Directory.CreateDirectory(Application.StartupPath + "\\TempExtract\\Data");
+                            Directory.CreateDirectory(Application.StartupPath + "\\TempExtract\\data\\models");
+                            InputPath = Application.StartupPath + "\\TempExtract\\data\\models\\" + Path.GetFileName(BuildPath.ToLower()).Substring(0, Path.GetFileName(BuildPath).Length - 3) + "map";
+                        }
+
+                        TrickyMirror.ProjectMirror(LoadPath.Substring(0, LoadPath.Length - 4), InputPath.Substring(0, InputPath.Length - 4));
+
+                        BigHandler bigHandler = new BigHandler();
+                        bigHandler.LoadFolderC0FB(Application.StartupPath + "\\TempExtract");
+                        bigHandler.CompressBuild = true;
+                        bigHandler.BuildBig(BuildPath);
+                    }
+                    else
+                    {
+                        TrickyMirror.ProjectMirror(LoadPath.Substring(0, LoadPath.Length - 4), InputPath.Substring(0, InputPath.Length - 4));
+                    }
+
+                    if (Directory.Exists(Application.StartupPath + "\\TempExtract"))
                     {
                         Directory.Delete(Application.StartupPath + "\\TempExtract", true);
                     }
